Add support for driving unselected row/col. (#16278)
