+++ keyboards/keychron/q2/config.h
@@ -29,6 +29,8 @@
 #define MATRIX_ROW_PINS { B4, B3, A15, A14, A13 }
 #define MATRIX_COL_PINS { C14, C15, A0, A1, A2, A3, A4, A5, A6, A7, B0, B1, A8, A9, H3 }
 
+#define MATRIX_UNSELECT_DRIVE_HIGH
+
 /* DIP switch */
 #define DIP_SWITCH_MATRIX_GRID  { {4, 4} }
